split conftest.py
one global conftest.py  and one for fully_connected networks
