+++ integration-testing/test/conftest.py
@@ -10,9 +10,9 @@ from tools.util import log_box
 from tools.profiling import log_prof_data
 import tools.resources as resources
 import os
-import collections
 import pprint
 from contextlib import contextmanager
+from tools.fixture import *
 
 def pytest_addoption(parser):
     parser.addoption(
@@ -31,21 +31,6 @@ def pytest_addoption(parser):
         "--rnode-timeout", action="store", default="10", help="timeout in seconds for executing an rnode call (Examples: propose, show-logs etc.). Defaults to 10s"
     )
 
-class RChain:
-    def __init__(self, network, bootstrap, peers):
-        self.network = network
-        self.bootstrap = bootstrap
-        self.peers = peers
-        self.nodes = [bootstrap] + peers
-
-Config = collections.namedtuple( "Config",
-                                 [
-                                     "peer_count",
-                                     "node_startup_timeout",
-                                     "network_converge_timeout",
-                                     "receive_timeout",
-                                     "rnode_timeout"
-                                 ])
 
 def parse_config(request):
     peer_count = int(request.config.getoption("--peer-count"))
@@ -100,7 +85,6 @@ def docker_network(docker):
             logging.info(f"removing {network.name}")
             network.remove()
 
-KeyPair = collections.namedtuple("KeyPair", ["private_key", "public_key"])
 
 @contextmanager
 def create_bonds_file(validator_keys):
@@ -150,34 +134,3 @@ def started_bootstrap(config, bootstrap):
               config.node_startup_timeout,
         "Bootstrap node didn't start correctly")
     yield bootstrap
-
-@pytest.fixture(scope="package")
-def rchain_network(config, docker, started_bootstrap, docker_network, validators_data):
-    logging.debug(f"Docker network = {docker_network}")
-
-    bonds_file, _, peer_keys = validators_data
-    peers = create_peer_nodes(docker, started_bootstrap, docker_network, bonds_file, peer_keys, config.rnode_timeout)
-
-    yield RChain(network = docker_network, bootstrap = started_bootstrap, peers = peers)
-
-    for peer in peers:
-        peer.cleanup()
-
-@pytest.fixture(scope="package")
-def started_rchain_network(config, rchain_network):
-    for peer in rchain_network.peers:
-        wait_for( string_contains(node_logs(peer),
-                                 "coop.rchain.node.NodeRuntime - Listening for traffic on rnode"),
-                  config.node_startup_timeout,
-                  f"Peer node {peer.name} didn't start correctly")
-
-    yield rchain_network
-
-
-@pytest.fixture(scope="package")
-def converged_network(config, started_rchain_network):
-    wait_for( network_converged( started_rchain_network.bootstrap, len(started_rchain_network.peers)),
-                                 config.network_converge_timeout,
-              "The network did NOT converge. Check container logs for issues. One or more containers might have failed to start or connect.")
-
-    yield started_rchain_network
